Some int to NSInteger/NSUInteger conversions
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3810 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/HelpViewer/Controller.m
@@ -70,7 +70,7 @@
 
 - (void) openFile: (id) sender
 {
-    int ret;
+    NSInteger ret;
     NSLog (@"openFile ...");
     NSOpenPanel* panel = [NSOpenPanel openPanel];
     [panel setAllowsMultipleSelection: NO];
